Bug 788021 followup: cast PR_NOW() to int64_t to fix a build failure on OpenBSD. r=jlebar
UltraBlame original commit: 4ff2eb1a97cdd6d98072d126d42692f67aab7e3b

+++ xpcom/base/nsMemoryReporterManager.cpp
@@ -1149,7 +1149,7 @@ nsMemoryReporterManager::DumpMemoryReportsToFile(
     
     nsString identifier(aIdentifier);
     if (identifier.IsEmpty()) {
-        identifier.AppendInt(PR_Now() / 1000000);
+        identifier.AppendInt(static_cast<int64_t>(PR_Now()) / 1000000);
     }
